backend/requirements.txt
# Core Flask
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
flask-jwt-extended==4.6.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# AI / ML
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
joblib==1.3.2

# Optional AI (comment out if not using XGBoost)
# xgboost==2.0.3
# shap==0.44.0

# Reports
openpyxl==3.1.2

# Background jobs (optional)
celery==5.3.6
redis==5.0.1
kombu==5.3.3

# Production server
gunicorn==21.2.0

# Utilities
python-dotenv==1.0.0
bcrypt==4.1.2
marshmallow==3.20.2
click==8.1.7

# Real-time WebSocket
flask-socketio==5.3.6
eventlet==0.35.2

# Background scheduler
apscheduler==3.10.4

# Rate limiting
flask-limiter==3.5.0

# AI extras
shap==0.44.0
flask-bcrypt==1.0.1
